requirements.txt
transformers
datasets
peft
accelerate
bitsandbytes
torch
gradio
tqdm
python-dotenv
PyGithub